refactored test code and scripts
