+++ jest.setup.ts
@@ -10,17 +10,21 @@ afterEach(() => {
 });
 //? Run react testing library's "cleanup" function after every test to prevent flow-on effects from consecutive renders
 
-Object.defineProperty(window, "matchMedia", {
-	writable: true,
-	value: jest.fn().mockImplementation((query) => ({
-		matches: false,
-		media: query,
-		onchange: null,
-		addListener: jest.fn(), // Deprecated
-		removeListener: jest.fn(), // Deprecated
-		addEventListener: jest.fn(),
-		removeEventListener: jest.fn(),
-		dispatchEvent: jest.fn(),
-	})),
-});
-//? We mock the "matchMedia" window method
+try {
+	Object.defineProperty(window, "matchMedia", {
+		writable: true,
+		value: jest.fn().mockImplementation((query) => ({
+			matches: false,
+			media: query,
+			onchange: null,
+			addListener: jest.fn(), // Deprecated
+			removeListener: jest.fn(), // Deprecated
+			addEventListener: jest.fn(),
+			removeEventListener: jest.fn(),
+			dispatchEvent: jest.fn(),
+		})),
+	});
+	//? We mock the "matchMedia" window method
+} catch (e) {
+	console.log("Skipped window property mock because of testing environment");
+}
