+++ package.json
@@ -7,8 +7,10 @@
 		"start:dev": "next dev",
 		"build": "next build",
 		"start": "next start",
-		"test": "jest",
+		"test": "yarn test:backend & yarn test:frontend",
 		"test:watch": "jest --watch",
+		"test:backend": "jest --testEnvironment=node --testPathPattern=/tests/backend",
+		"test:frontend": "jest --testPathPattern=/tests/frontend",
 		"setupDevDb": "ts-node scripts/setupDevDb.ts"
 	},
 	"dependencies": {
